skb: move event to events mod
