+++ src/events/mod.rs
@@ -20,3 +20,5 @@ pub(crate) use display::*;
 
 pub(crate) mod bpf;
 pub(crate) mod file;
+
+pub(crate) mod skb;
