+++ src/module/skb/mod.rs
@@ -7,9 +7,6 @@
 pub(crate) mod skb;
 pub(crate) use skb::*;
 
-pub(crate) mod event;
-pub(crate) use event::*;
-
 mod bpf;
 mod skb_hook {
     include!("bpf/.out/skb_hook.rs");
